+++ src/configs/index.ts
@@ -6,6 +6,7 @@ import {
 export const baseUrl = 'http://localhost:3000'
 export const apiUrl = 'https://escrow-api.linkdrop.io/v3'
 export const dashboardApiUrl = 'https://escrow-api.linkdrop.io/dashboard'
+export const devDashboardApiUrl = 'https://escrow-api.linkdrop.io/dev-dashboard'
 
 export const nativeTokenAddress = '0x0000000000000000000000000000000000000000'
 
@@ -14,6 +15,10 @@ export const mainEscrowContract = '0x88d51990a3b962f975846f3688e36d2a1fc611f1'
 
 export const cbwEscrowContractNFT = '0xe07fa88a10a915b7339aff050db82c0030bf6861'
 export const mainEscrowContractNFT = '0x648b9a6c54890a8fb17de128c6352f621154f358'
+
+export const immutableZkevmContract = '0x4366caf3963d147da4a4287061354058d871d1be'
+export const immutableZkevmContractNFT = '0x317d2501396fe75d997799bf3bdbc7cc6768b533'
+
 
 export const supportedStableCoins = {
   [ETokenAddress.usdcBase]: ESelectors.receiveWithAuthorization,