+++ src/helpers/define-if-escrow-address-is-correct.ts
@@ -1,17 +1,19 @@
 import * as configs from '../configs'
-import { TDeploymentType, TTokenType } from '../types'
+import { EChains, TDeploymentType, TTokenType } from '../types'
 import defineVersionByEscrow from './define-version-by-escrow'
 
 type TDefineIfEscrowAddressIsCorrect = (
   escrowAddress: string,
   tokenType: TTokenType,
-  deployment: TDeploymentType
+  deployment: TDeploymentType,
+  chainId: number | null
 ) => boolean
 
 const defineIfEscrowAddressIsCorrect: TDefineIfEscrowAddressIsCorrect = (
   escrowAddress,
   tokenType,
-  deployment
+  deployment,
+  chainId
 ) => {
 
   if (!escrowAddress || !tokenType) {
@@ -40,6 +42,13 @@ const defineIfEscrowAddressIsCorrect: TDefineIfEscrowAddressIsCorrect = (
       return false
     }
     return true
+  }
+
+  if (chainId === EChains.immutableZkevm) {
+    if (tokenType === 'ERC1155' || tokenType === 'ERC721') {
+      return escrowAddress === configs.immutableZkevmContractNFT
+    }
+    return escrowAddress === configs.immutableZkevmContract
   }
 
   if (tokenType === 'ERC1155' || tokenType === 'ERC721') {